Fix semester status change
